+++ js/zoomhome.js
@@ -252,7 +252,10 @@ function leafext_zoomhome_js(maps,map_id,allfit,position) {
 		}
 	);
 
-	// maps[map_id].on("zoomend", function(e) {
-	// console.log("zoomend zoom "+map_id+" "+maps[map_id].getZoom());
-	// });
+	// maps[map_id].on(
+	// 	"zoomend",
+	// 	function (e) {
+	// 		console.log( "zoomend zoom " + map_id + " " + maps[map_id].getZoom() );
+	// 	}
+	// );
 }